+++ lib/analysis/buffer_observer.dart
